moved train_utils to a logical place
