+++ src/unexploitable_search/mitigation_strats/train.py
@@ -10,9 +10,7 @@ from transformers import AutoModelForCausalLM
 import wandb
 from unexploitable_search.callbacks.best_reward import SaveBestRewardCallback
 from unexploitable_search.config.core import PROJECT_ROOT
-from unexploitable_search.model_organisms.train_utils import (
-    make_quest_grpo_reward,
-)
+from unexploitable_search.train_utils import make_quest_grpo_reward
 from unexploitable_search.settings.apps.sandbox import DockerSandbox
 from unexploitable_search.settings.apps.setting import APPSSetting
 from unexploitable_search.trainers.maxent_trainer import (
@@ -99,11 +97,13 @@ def main(cfg: DictConfig):
 
     args = instantiate(cfg.train.trl_trainer_config)
 
+    # specify the sandbox in which the APPS reward function should run unit tests
     additional_kwargs = {}
     if isinstance(setting, APPSSetting):
         additional_kwargs = {
             "sandbox": DockerSandbox(),
         }
+
     reward_funcs = [make_quest_grpo_reward(setting.main_quest, additional_kwargs)]
 
     if cfg.mitigation_strat is None:
