+++ src/unexploitable_search/model_organisms/train.py
@@ -11,9 +11,7 @@ from trl.trainer.grpo_trainer import GRPOTrainer
 import wandb
 from unexploitable_search.callbacks.best_reward import SaveBestRewardCallback
 from unexploitable_search.config.core import PROJECT_ROOT
-from unexploitable_search.model_organisms.train_utils import (
-    make_quest_grpo_reward,
-)
+from unexploitable_search.train_utils import make_quest_grpo_reward
 from unexploitable_search.quest import Quest
 from unexploitable_search.setting import Setting
